requirements.txt
fastapi
pydantic
requests
python-dotenv
uvicorn

# Testes
pytest
pytest-cov
pytest-mock
pytest-asyncio
